+++ src/State/CritterInteract.cpp
@@ -25,6 +25,7 @@
 #include <memory>
 
 // Falltergeist includes
+#include "../Audio/IMixer.h"
 #include "../Format/Lst/File.h"
 #include "../Format/Lip/File.h"
 #include "../Game/CritterObject.h"
@@ -43,7 +44,6 @@
 #include "../UI/Animation.h"
 #include "../UI/AnimationFrame.h"
 #include "../UI/AnimationQueue.h"
-#include "../Audio/Mixer.h"
 
 // Third party includes
 
@@ -73,16 +73,14 @@ namespace Falltergeist
         void CritterInteract::onStateActivate(Event::State* event)
         {
             Game::getInstance()->mouse()->pushState(Input::Mouse::Cursor::BIG_ARROW);
-            if (_headID >= 0)
-            {
+            if (_headID >= 0) {
                 // stop music completely
                 // TODO: because Dialog state is activated *before* Interact, this stops speech too =/
                 //Game::getInstance()->mixer()->stopMusic();
-            }
-            else
-            {
+            } else {
                 // lower music volume
-                Game::getInstance()->mixer()->setMusicVolume(Game::getInstance()->mixer()->musicVolume()/2.0);
+                auto mixer = Game::getInstance()->mixer();
+                mixer->setChannelVolume(Audio::Channel::Music, mixer->channelVolume(Audio::Channel::Music) / 2.0);
             }
         }
 
@@ -96,7 +94,8 @@ namespace Falltergeist
             else
             {
                 // restore music volume
-                Game::getInstance()->mixer()->setMusicVolume(Game::getInstance()->mixer()->musicVolume()*2.0);
+                auto mixer = Game::getInstance()->mixer();
+                mixer->setChannelVolume(Audio::Channel::Music, mixer->channelVolume(Audio::Channel::Music) * 2.0);
             }
         }
 
@@ -329,7 +328,7 @@ namespace Falltergeist
 
         void CritterInteract::switchSubState(CritterInteract::SubState state)
         {
-            Game::getInstance()->mixer()->stopMusic();
+            Game::getInstance()->mixer()->stopChannel(Audio::Channel::Music);
             _phase = Phase::FIDGET;
             _fidgetTimer.start(0);
             if (_state!=SubState::NONE)
@@ -356,7 +355,7 @@ namespace Falltergeist
 
         void CritterInteract::transition(Reaction reaction)
         {
-            Game::getInstance()->mixer()->stopMusic();
+            Game::getInstance()->mixer()->stopChannel(Audio::Channel::Music);
             auto newmood = _mood;
 
             if (headID()!= -1)
@@ -453,8 +452,4 @@ namespace Falltergeist
             script()->run();
         }
     }
-
-
-
-
 }