+++ src/State/Movie.cpp
@@ -23,7 +23,7 @@
 // C++ standard includes
 
 // Falltergeist includes
-#include "../Audio/Mixer.h"
+#include "../Audio/IMixer.h"
 #include "../CrossPlatform.h"
 #include "../Event/Keyboard.h"
 #include "../Event/Mouse.h"
@@ -183,7 +183,7 @@ namespace Falltergeist
 
         void Movie::onVideoFinished()
         {
-            Game::getInstance()->mixer()->stopMusic();
+            Game::getInstance()->mixer()->stopChannel(Audio::Channel::Music);
             Game::getInstance()->mouse()->popState();
 
             // without this the location will be absolutely dark if you didn't interrupted movie play