+++ src/UI/ImageButton.cpp
@@ -24,7 +24,7 @@
 #include <string>
 
 // Falltergeist includes
-#include "../Audio/Mixer.h"
+#include "../Audio/IMixer.h"
 #include "../Event/Event.h"
 #include "../Event/Mouse.h"
 #include "../Exception.h"